Add packaging to sqlitefid
A number of chnges are made to make sqlitefid easy to package and for
it to be distributed through the standard Python packaging mechanism.
The code layout is improved. A Makefile has been added to perform the
packaging and upload. Imports have been fixed for the various
different ways that the code can be run, i.e. from the repository
root or installed as an executable.