Add guard for fnames that terminate incorrectly
Spotted in the wild recently, filenames that include a 0x0D character
which cause problems with analysis. These filenames should terminate
in YAML with (') and so if we cannot detect one, we need to read the
rext line and add it to the first line. The report will then generate
correctly.